Promote TreehouseApp.Binding to a top-level file
I'm refactoring in TreehouseApp and it's doing too many things.
